Update merge commit SHA fetching flow
The earlier implementation wen through a list of commits in the branches. This would've been inefficient since it will traverse through a lot of commits if the revert is requested past a certain amount of days.

So I moved to using the PR itslef to find the merge commit and the PR body and title

Added some code imporvements